Split main file into multiple dedicated files
