+++ src/events/quit_listener.go
@@ -1,0 +1,10 @@
+package events
+
+import (
+	"os"
+)
+
+func QuitListener(interruptChan chan os.Signal, doneChan chan bool) {
+	<-interruptChan
+	doneChan <- false
+}
