+++ src/game/game.go
@@ -4,7 +4,6 @@ import (
 	"errors"
 	"fmt"
 	"reflect"
-	"unicode"
 )
 
 type Grid [][]byte
