+++ src/renderer/terminal.go
@@ -1,0 +1,12 @@
+package renderer
+
+import (
+	"os"
+	"os/exec"	
+)
+
+func ClearTerminal() {
+	clearCmd := exec.Command("clear")
+	clearCmd.Stdout = os.Stdout
+	clearCmd.Run()
+}
